renamed from cxx_client to db
